Fix survival remaining hours calculation

+++ HowLongToBeatSteam/Controllers/Responses/RemainingLifeExpectancyResponse.cs
@@ -9,7 +9,7 @@ namespace HowLongToBeatSteam.Controllers.Responses
         public string date { get; set; }
         public string country { get; set; }
         public string age { get; set; }
-        public float remaining_life_expectancy { get; set; }
+        public double remaining_life_expectancy { get; set; }
         public string sex { get; set; }
     }
     // ReSharper restore InconsistentNaming
